feat(#201): Home 영역 코드 정리
