--- src/components/home/journals/shortform/ShortForm/index.tsx
@@ -1,376 +1,0 @@
-import YouTube from 'react-youtube';
-import { useNavigate } from 'react-router-dom';
-import toast from 'react-hot-toast';
-import { Suspense, useEffect, useRef, useState } from 'react';
-
-import UserIcon from '@_icons/user.svg?react';
-import LocationIcon from '@_icons/location.svg?react';
-import ThemeIcon from '@_icons/theme.svg?react';
-import BookMarkIcon from '@_icons/bookmark.svg?react';
-import FilledBookMarkIcon from '@_icons/bookmark_filled_white.svg?react';
-import CommentIcon from '@_icons/comment.svg?react';
-import LikeIcon from '@_icons/clap.svg?react';
-import FilledLikeIcon from '@_icons/clap_blue.svg?react';
-import PlaceIcon from '@_icons/place.svg?react';
-import ShareIcon from '@_icons/share.svg?react';
-
-import Typography from '@_common/Typography';
-import { Toast } from '@_common/Toast';
-import { get, post } from '@_utils/api';
-import useModal from '../../../../../hooks/useModal';
-import useAlert from '../../../../../hooks/useAlert';
-import usePopup from '../../../../../hooks/usePopup';
-
-import useScrapModal from '../../../../video/useScrapModal';
-import Comment from '../../../../journal/Comment';
-
-import PlacesModalContents from '../PlacesModalContents';
-import * as S from './style';
-
-export interface TShortForm {
-  id: number;
-  title: string;
-  location: string[];
-  theme: string[];
-  place: string[];
-  createdAt: string;
-  videoId: string;
-  content: string;
-  userid: string;
-  username: string;
-  profileImage: string;
-  clap: number;
-  isClapped: boolean;
-  bookmark: number;
-  isBookmarked: boolean;
-  commentCount: number;
-}
-
-export interface Props {
-  shortFormId: number;
-  visible: boolean;
-  videoId: string;
-}
-
-function ShortForm({ shortFormId, visible, videoId }: Props) {
-  const [data, setData] = useState<TShortForm>();
-  // @todo: 스크랩 데이터에서 스크랩됐는지 비교 => 북마크 활성화/비활성화 관리
-  const [clap, setClap] = useState<{
-    count: number;
-    isActive: boolean;
-  }>({ count: 0, isActive: false });
-  const [bookmark, setBookmark] = useState<{
-    count: number;
-    isActive: boolean;
-  }>({ count: 0, isActive: false });
-  const [contentsCommentCount, setContentsCommentCount] = useState<number>(0);
-
-  const containerRef = useRef<HTMLDivElement>(null);
-  const loadingRef = useRef<HTMLDivElement>(null);
-  const contentContainerRef = useRef<HTMLParagraphElement>(null);
-  const playerRef = useRef<YouTube>(null);
-  const [isCaptionOpened, setIsCaptionOpened] = useState<boolean>(false);
-  const { Modal: PlacesModal, modalOpen: placesModalOpen } = useModal({
-    title: '태그된 장소',
-    handle: true,
-  });
-  const { ScrapModal, scrapModalOpen } = useScrapModal({
-    id: shortFormId,
-    type: 'short-form',
-    setIsScraped: () => {
-      setBookmark((prev) => ({
-        count: prev.isActive ? prev.count - 1 : prev.count + 1,
-        isActive: !prev.isActive,
-      }));
-    },
-  });
-  const navigate = useNavigate();
-  const { Alert: LoginALlert, alertOpen: loginAlertOpen } = useAlert({
-    Content: (
-      <Typography.Body size="lg" color="white">
-        로그인이 필요한 서비스에요.
-      </Typography.Body>
-    ),
-    RightContent: (
-      <Typography.Body size="lg" color="white">
-        <span
-          style={{ textDecoration: 'underline', cursor: 'pointer' }}
-          onClick={() => {
-            navigate('/login');
-          }}
-        >
-          로그인 하러가기
-        </span>
-      </Typography.Body>
-    ),
-  });
-
-  const toggleClap = async () => {
-    post<{ community: string; postId: number }>('/clap/community', {
-      community: 'short-form',
-      postId: shortFormId,
-    });
-    if (clap.isActive) {
-      setClap((prev) => ({ count: prev.count - 1, isActive: false }));
-    } else {
-      setClap((prev) => ({ count: prev.count + 1, isActive: true }));
-    }
-  };
-
-  const videoPlayControl = () => {
-    if (playerRef.current === null) return;
-    const player = playerRef.current?.getInternalPlayer();
-    if (visible) {
-      player.playVideo();
-    } else {
-      player.pauseVideo();
-    }
-  };
-
-  const { Modal, modalOpen } = useModal({
-    title: '',
-    handle: false,
-    borderRadius: '16px',
-  });
-  const { Popup, popupOpen, popupClose } = usePopup();
-
-  const createCopyURL = (id: number) => {
-    const arr = window.location.href.split('/').slice(0, -1);
-    arr.push(String(id));
-    return arr.join('/');
-  };
-
-  useEffect(() => {
-    videoPlayControl();
-  }, [visible]);
-
-  useEffect(() => {
-    if (!isCaptionOpened && contentContainerRef.current !== null) {
-      contentContainerRef.current.scrollTop = 0;
-    }
-  }, [isCaptionOpened]);
-
-  useEffect(() => {
-    const { current: container } = containerRef;
-    playerRef.current?.getInternalPlayer().setSize('100%', container?.offsetHeight);
-  }, [data]);
-
-  useEffect(() => {
-    const options = {
-      root: null,
-      rootMargin: '0px',
-      threshold: 0.1,
-    };
-
-    const observer = new IntersectionObserver((entries) => {
-      entries.forEach((entry) => {
-        if (entry.isIntersecting) {
-          const fetchData = async () => {
-            const { data } = await get<TShortForm>(`/community/short-form/${shortFormId}`);
-            setData(data);
-            setContentsCommentCount(data.commentCount);
-            setClap({ count: data.clap, isActive: data.isClapped });
-            setBookmark({ count: data.bookmark, isActive: data.isBookmarked });
-          };
-          fetchData();
-        }
-      });
-    }, options);
-
-    if (loadingRef.current) {
-      observer.observe(loadingRef.current);
-    }
-
-    return () => observer.disconnect();
-  }, []);
-
-  return (
-    <>
-      {data && (
-        <>
-          <LoginALlert />
-          <PlacesModal>
-            <PlacesModalContents id={shortFormId} />
-          </PlacesModal>
-          <Modal>
-            <Comment
-              id={shortFormId}
-              commentInputPosition="bottom"
-              type="short-form"
-              commentCount={data.commentCount}
-              setContentsCommentCount={setContentsCommentCount}
-            />
-          </Modal>
-          <Popup>
-            <S.UrlLabel htmlFor="urlCopy">
-              <Typography.Title size="md">아래 링크를 복사해 공유해보세요!</Typography.Title>
-            </S.UrlLabel>
-            <S.UrlInput
-              type="url"
-              name="현재 링크 복사"
-              id="urlCopy"
-              value={createCopyURL(shortFormId)}
-              onClick={(e) => {
-                e.preventDefault();
-                navigator.clipboard.writeText(`${data.title}\n${createCopyURL(shortFormId)}`);
-                toast.custom(() => (
-                  <Toast>
-                    <Typography.Title size="md" color="white">
-                      URL이 복사되었습니다.
-                    </Typography.Title>
-                  </Toast>
-                ));
-                popupClose();
-              }}
-            />
-          </Popup>
-          <ScrapModal />
-          <S.Container>
-            <S.YoutubeContainer isCaptionOpened={isCaptionOpened} ref={containerRef}>
-              <Suspense fallback={<S.YoutubeFallback>로딩 중...</S.YoutubeFallback>}>
-                <YouTube
-                  videoId={videoId}
-                  ref={playerRef}
-                  loading="lazy"
-                  opts={{
-                    width: '100%',
-                    height: '100px',
-                    playerVars: {
-                      autoplay: 1,
-                      controls: 0,
-                      showInfo: 0,
-                      rel: 0,
-                      modestbranding: 0,
-                      loop: 1,
-                      autohide: 1,
-                      playlist: videoId,
-                    },
-                  }}
-                  onEnd={() => {
-                    playerRef.current?.resetPlayer();
-                  }}
-                  onReady={() => {
-                    const player = playerRef.current?.getInternalPlayer();
-                    player.setVolume(40);
-
-                    videoPlayControl();
-                  }}
-                />
-              </Suspense>
-            </S.YoutubeContainer>
-            <S.InfoBox>
-              {/* 업로드 유저 프로필 */}
-              <p>
-                <S.ProgileImageBox>
-                  {data.profileImage ? <img src={data.profileImage} alt="" /> : <UserIcon />}
-                </S.ProgileImageBox>
-                <span>
-                  <Typography.Body size="lg" color="white">
-                    {data.username}
-                  </Typography.Body>
-                </span>
-              </p>
-              <p>{data.title}</p>
-              {/* 제목 아래 캡션 영역 */}
-              <S.ContentBox
-                ref={contentContainerRef}
-                isOpened={isCaptionOpened}
-                onClick={() => {
-                  setIsCaptionOpened((prev) => !prev);
-                }}
-              >
-                {isCaptionOpened ? (
-                  data.content.split('\n').map((line) => <p>{line}</p>)
-                ) : (
-                  <p>{data.content.split('\n')[0]}</p>
-                )}
-              </S.ContentBox>
-              <S.BottomInfoContainer>
-                <span>
-                  <LocationIcon />
-                  {data.location}
-                </span>
-                {data.theme.length > 0 && (
-                  <span>
-                    <ThemeIcon />
-                    {data.theme.join(', ')}
-                  </span>
-                )}
-              </S.BottomInfoContainer>
-            </S.InfoBox>
-            <S.ControlBox>
-              <S.IconButton
-                onClick={() => {
-                  if (localStorage.getItem('access_token')) {
-                    toggleClap();
-                  } else {
-                    loginAlertOpen();
-                  }
-                }}
-              >
-                {clap.isActive ? <FilledLikeIcon /> : <LikeIcon />}
-                {clap.count}
-              </S.IconButton>
-              <S.IconButton
-                onClick={() => {
-                  if (localStorage.getItem('access_token')) {
-                    modalOpen();
-                  } else {
-                    loginAlertOpen();
-                  }
-                }}
-              >
-                <CommentIcon />
-                {contentsCommentCount}
-              </S.IconButton>
-              <S.IconButton
-                onClick={(e) => {
-                  e.preventDefault();
-                  if (localStorage.getItem('access_token')) {
-                    if (!bookmark.isActive) {
-                      post<{ message: 'Create Success' | 'Delete Success' }>(
-                        '/folder/scrap/community',
-                        {
-                          community: 'short-form',
-                          postId: shortFormId,
-                        },
-                      ).then(() => {
-                        setBookmark((prev) => ({
-                          count: prev.count + 1,
-                          isActive: true,
-                        }));
-                      });
-                    }
-                    scrapModalOpen();
-                  } else {
-                    loginAlertOpen();
-                  }
-                }}
-              >
-                {bookmark.isActive ? <FilledBookMarkIcon /> : <BookMarkIcon />}
-                {bookmark.count}
-              </S.IconButton>
-              <S.IconButton
-                onClick={() => {
-                  if (localStorage.getItem('access_token')) {
-                    placesModalOpen();
-                  } else {
-                    loginAlertOpen();
-                  }
-                }}
-              >
-                <PlaceIcon />
-              </S.IconButton>
-              <S.IconButton onClick={popupOpen}>
-                <ShareIcon />
-              </S.IconButton>
-            </S.ControlBox>
-          </S.Container>
-        </>
-      )}
-      {!data && <div ref={loadingRef} style={{ height: '100dvh' }} />}
-    </>
-  );
-}
-
-export default ShortForm;
