+++ src/pages/Article/ArticlesPage/style.ts
@@ -1,4 +1,4 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 export const Container = styled.div`
   padding-top: 20px;
