+++ src/pages/Home/index.tsx
@@ -1,24 +1,16 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useRecoilState } from 'recoil';
-import { useNavigate, useSearchParams, Link } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 import PageTemplate from '../../components/common/PageTemplate';
-import TabBar from '../../components/common/TabBar';
-
-import Journals from '../../components/home/journals/Journals';
-import Recommendation from '../../components/home/Recommendation';
-import Articles from '../../components/home/Articles';
 
 import * as S from './style';
 import useAlert from '../../hooks/useAlert';
 import Typography from '../../components/common/Typography';
 import { loginAlertState } from '../../recoil/loginAlertState';
+import Recommendation from '../../components/home/Recommendation';
 
 function HomePage() {
-  const [query] = useSearchParams();
-  const [focusedTabIndex, setFocusedTabIndex] = useState<number>(
-    query.get('tab') ? Number(query.get('tab')) : 0,
-  );
   const navigate = useNavigate();
   const [isLoginAlertState, setIsLoginAlertState] = useRecoilState(loginAlertState);
   const { Alert, alertOpen } = useAlert({
@@ -40,11 +32,6 @@ function HomePage() {
       </Typography.Body>
     ),
   });
-  // const tabs = [
-  //   { id: '추천', name: '추천', content: <Recommendation /> },
-  //   { id: '아티클', name: '아티클', content: <Articles /> },
-  //   { id: '숏폼', name: '숏폼', content: <Journals /> },
-  // ];
 
   useEffect(() => {
     if (isLoginAlertState) {
@@ -54,22 +41,8 @@ function HomePage() {
   }, []);
 
   return (
-    <PageTemplate
-      header={
-        <S.Header>
-          {/* <TabBar
-            tabs={tabs}
-            focusedTabIndex={focusedTabIndex}
-            setFocusedTabIndex={setFocusedTabIndex}
-            widthStyle="fit-content"
-            fontSize="20px"
-            color="#424242"
-          /> */}
-        </S.Header>
-      }
-    >
+    <PageTemplate>
       <Alert />
-      {/* {tabs[focusedTabIndex].content} */}
       <Recommendation />
     </PageTemplate>
   );
